+++ actix-cors/src/error.rs
@@ -2,6 +2,8 @@ use actix_web::{http::StatusCode, HttpResponse, ResponseError};
 
 use derive_more::{Display, Error};
 
+use crate::inner::add_vary_header;
+
 /// Errors that can occur when processing CORS guarded requests.
 #[derive(Debug, Clone, Display, Error)]
 #[non_exhaustive]
@@ -45,6 +47,8 @@ impl ResponseError for CorsError {
     }
 
     fn error_response(&self) -> HttpResponse {
-        HttpResponse::with_body(StatusCode::BAD_REQUEST, self.to_string()).map_into_boxed_body()
+        let mut res = HttpResponse::with_body(self.status_code(), self.to_string());
+        add_vary_header(res.headers_mut());
+        res.map_into_boxed_body()
     }
 }