+++ actix-cors/src/middleware.rs
@@ -1,4 +1,4 @@
-use std::{collections::HashSet, convert::TryInto, rc::Rc};
+use std::{collections::HashSet, rc::Rc};
 
 use actix_utils::future::ok;
 use actix_web::{
@@ -13,7 +13,7 @@ use actix_web::{
 use futures_util::future::{FutureExt as _, LocalBoxFuture};
 use log::debug;
 
-use crate::{builder::intersperse_header_values, AllOrSome, Inner};
+use crate::{builder::intersperse_header_values, inner::add_vary_header, AllOrSome, Inner};
 
 /// Service wrapper for Cross-Origin Resource Sharing support.
 ///
@@ -67,7 +67,9 @@ impl<S> CorsMiddleware<S> {
             res.insert_header((header::ACCESS_CONTROL_MAX_AGE, max_age.to_string()));
         }
 
-        let res = res.finish();
+        let mut res = res.finish();
+        add_vary_header(res.headers_mut());
+
         req.into_response(res)
     }
 
@@ -116,21 +118,7 @@ impl<S> CorsMiddleware<S> {
         }
 
         if inner.vary_header {
-            let value = match res.headers_mut().get(header::VARY) {
-                Some(hdr) => {
-                    let mut val: Vec<u8> = Vec::with_capacity(hdr.len() + 71);
-                    val.extend(hdr.as_bytes());
-                    val.extend(
-                        b", Origin, Access-Control-Request-Method, Access-Control-Request-Headers",
-                    );
-                    val.try_into().unwrap()
-                }
-                None => HeaderValue::from_static(
-                    "Origin, Access-Control-Request-Method, Access-Control-Request-Headers",
-                ),
-            };
-
-            res.headers_mut().insert(header::VARY, value);
+            add_vary_header(res.headers_mut());
         }
 
         res
@@ -143,7 +131,6 @@ where
     S::Future: 'static,
 
     B: MessageBody + 'static,
-    B::Error: Into<Error>,
 {
     type Response = ServiceResponse<EitherBody<B>>;
     type Error = Error;
@@ -173,12 +160,7 @@ where
             async move {
                 let res = fut.await;
 
-                if origin.is_some() {
-                    Ok(Self::augment_response(&inner, res?))
-                } else {
-                    res.map_err(Into::into)
-                }
-                .map(|res| res.map_into_left_body())
+                Ok(Self::augment_response(&inner, res?).map_into_left_body())
             }
             .boxed_local()
         }
@@ -189,13 +171,20 @@ where
 mod tests {
     use actix_web::{
         dev::Transform,
+        middleware::Compat,
         test::{self, TestRequest},
+        App,
     };
 
     use super::*;
     use crate::Cors;
 
-    #[actix_rt::test]
+    #[test]
+    fn compat_compat() {
+        let _ = App::new().wrap(Compat::new(Cors::default()));
+    }
+
+    #[actix_web::test]
     async fn test_options_no_origin() {
         // Tests case where allowed_origins is All but there are validate functions to run incase.
         // In this case, origins are only allowed when the DNT header is sent.