+++ actix-session/examples/basic.rs
@@ -1,8 +1,6 @@
 use actix_session::{storage::RedisActorSessionStore, Session, SessionMiddleware};
 use actix_web::cookie::Key;
 use actix_web::{middleware, web, App, Error, HttpRequest, HttpServer, Responder};
-use rand::distributions::Alphanumeric;
-use rand::{thread_rng, Rng};
 
 /// simple handler
 async fn index(req: HttpRequest, session: Session) -> Result<impl Responder, Error> {
@@ -19,21 +17,12 @@ async fn index(req: HttpRequest, session: Session) -> Result<impl Responder, Err
     Ok("Welcome!")
 }
 
-// The signing key would usually be read from a configuration file/environment variables.
-fn get_signing_key() -> Key {
-    let signing_key: String = thread_rng()
-        .sample_iter(&Alphanumeric)
-        .take(64)
-        .map(char::from)
-        .collect();
-    Key::from(signing_key.as_bytes())
-}
-
-#[actix_rt::main]
+#[actix_web::main]
 async fn main() -> std::io::Result<()> {
     std::env::set_var("RUST_LOG", "actix_web=info,actix_redis=info");
     env_logger::init();
-    let signing_key = get_signing_key();
+    // The signing key would usually be read from a configuration file/environment variables.
+    let signing_key = Key::generate();
 
     HttpServer::new(move || {
         App::new()